Move uploads controller to controllers/
